Added Run Experiment Method for Benchmarking
